Refactor HIDUsage
